refactor(domain, view, controller): step5 리뷰 반영
- util 패키지를 view로 변경
- strategy 패턴이므로 dto 로 생각했던 부분들은 다 domain으롭 ㅕㄴ경
- 매번 new RandomValue() 객체 생성하는 것 수정 -> 매번 객체생성하는 부분이 있다면 항상 생각하기
- 비교연산자 자체로 true, false 반환하므로 return 간단하게 수
