+++ src/main/java/racingcar/domain/Racing.java
@@ -2,6 +2,7 @@ package racingcar.domain;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.Random;
 import java.util.stream.Collectors;
 
 public class Racing {
@@ -24,7 +25,7 @@ public class Racing {
     }
 
     public CarGroup runRacing(){
-        carGroup.moveAll();
+        carGroup.moveAll(new RandomValue());
         return carGroup;
     }
