+++ src/test/java/racingcar/CarTest.java
@@ -3,9 +3,9 @@ package racingcar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import racingcar.domain.Car;
-import racingcar.dto.MovableRandomValueGenerator;
-import racingcar.dto.NonMovableRandomValueGenerator;
-import racingcar.dto.RandomValueGenerator;
+import racingcar.domain.MovableRandomValueGenerator;
+import racingcar.domain.NonMovableRandomValueGenerator;
+import racingcar.domain.RandomValueGenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
